requirements.txt
astroid==2.4.2
autopep8==1.5.4
branca==0.3.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
Django==2.2.16
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
folium==0.10.1
idna==2.8
isort==4.3.21
Jinja2==2.10.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
numpy==1.17.3
pandas==0.25.3
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
scipy==1.3.1
seaborn==0.9.0
six==1.13.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.7
wrapt==1.12.1
